requirements.txt
streamlit>=1.35
plotly>=5.0
python-dotenv>=1.0
pandas>=1.5
numpy>=1.23
requests>=2.28
